requirements.txt
fastapi[standard]
uvicorn
jinja2
mysql-connector-python
python-multipart
python-dotenv
google-genai
google.generativeai
